LPARAM replaced with nint

+++ Au/Util/ARecording.cs
@@ -25,8 +25,8 @@ namespace Au.Util
 		/// </summary>
 		/// <param name="recorded">
 		/// List of x y distances from previous.
-		/// The first distance is from the mouse position before the first movement; at run time it will be distance from <see cref="AMouse.LastXY"/>.
-		/// To create uint value from distance dx dy use this code: <c>AMath.MakeUint(dx, dy)</c>.
+		/// The first distance is from mouse position before the first movement; at run time it will be distance from <see cref="AMouse.LastXY"/>.
+		/// To create uint value from distance dx dy use <see cref="AMath.MakeLparam"/> and cast to uint.
 		/// </param>
 		/// <param name="withSleepTimes">
 		/// <i>recorded</i> also contains sleep times (milliseconds) alternating with distances.
@@ -55,8 +55,8 @@ namespace Au.Util
 					//never mind: ~90% is 7. Removing it would make almost 2 times smaller string. But need much more code. Or compress (see comment below).
 				} else {
 					//info: to make more compact, we write not distances (dx dy) but distance changes (x y).
-					int dx = AMath.LoShort(u), x = dx - pdx; pdx = dx;
-					int dy = AMath.HiShort(u), y = dy - pdy; pdy = dy;
+					int dx = AMath.LoShort((nint)u), x = dx - pdx; pdx = dx;
+					int dy = AMath.HiShort((nint)u), y = dy - pdy; pdy = dy;
 
 					if(x >= -4 && x < 4 && y >= -4 && y < 4) nbytes = 1; //3+3+2=8 bits, 90%
 					else if(x >= -64 && x < 64 && y >= -64 && y < 64) nbytes = 2; //7+7+2=16 bits, ~10%
